+++ read_pcap.py
@@ -151,13 +151,7 @@ class BeamformReader:
         return [self.read_bfee(x) for x in frames]
 
 if __name__ == "__main__":
-
-    if len(sys.argv) > 1:
-        path = sys.argv[1]
-    else:
-        # basePath = Path(__file__).parent
-        # path = (basePath / "../../sample_data/out.pcap").resolve()
-        path = r"E:\\DataLab PhD Albyn 2018\\Code\\sample_data\\out.pcap"
+    path = ".\\data\\pi\\walk_1597159475.pcap"
 
     reader = BeamformReader(path, "43455c0")
